datapipeline/requirements.txt
apache-airflow==2.7.0
apache-airflow-providers-common-sql==1.7.1
apache-airflow-providers-ftp==3.5.1
apache-airflow-providers-http==4.5.1
apache-airflow-providers-imap==3.3.1
apache-airflow-providers-sqlite==3.4.3
kafka-python>=2.0.2
requests>=2.31.0
pandas>=2.0.3
numpy>=1.24.4
pyarrow>=12.0.1